chore(GitHub): Add RFC workflow and release notes documentation (#45)
Since this repo now has multiple active consumers, it seems like a good time to raise the level of transparency for incoming changes. Not only will this help consumers and contributors keep up in an asynchronous way, but it will also provide a more structured way for active contributors to elaborate on their ideas, gain consensus and get actionable feedback.