fix(prettier): fix Prettier config issue and update formatting
Since our Prettier config file was called ‘prettier.config.js’, it was being picked up by the latest version of Prettier and throwing an error because it was inside of an empty folder. I’ve renamed it to ‘prettierConfig.js’ to fix the issue, but also renamed the other config files to match this new convention.

The latest version of Prettier has some formatting tweaks, so I’ve also reformatted the codebase.