fix(postinstall): Centralise handling of cwd to support postinstall (#198)
When `sku configure` is run by the `postinstall` hook, `process.cwd()` is `path/to/node_modules/sku`. This causes the config files controlled by the `configure` script to be written into `node_modules` instead of the project root.

From [the docs](https://docs.npmjs.com/cli/run-script):
> [npm] scripts are run from the root of the module [which defines them], regardless of what your current working directory is when you call npm run. If you want your script to use different behaviour based on what subdirectory you're in, you can use the INIT_CWD environment variable, which holds the full path you were in when you ran npm run.

npm provides our desired path as the `INIT_CWD` environment variable, so the core of this change is a `cwd` function that centralises this logic.